Add a skeleton gamepad resource.
This implements the skeleton of the gamepad resource for the IPC proxy. It is not actually hooked up. Hooking it up will require moving some gamepad lock code to a shared location.

This also hooks up the browser message routing for implementing resource hosts in the browser process.

BUG=

Review URL: https://chromiumcodereview.appspot.com/10824272

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@153265 0039d316-1c4b-4281-b951-d872f2087c98