Revert 153933 - Make history easily filterable.
This breaks tests that use hand rolled paths instead of the main path methods.

We're renaming the two profile directories that we want to keep, and dropping
GTK specific tests in anticipation of the changeover later this year.

Every change to the history, theme or profile format required a new giant
binary file added to our repository. There are tens of versions of the history
database alone, each weighing at over 10 megabytes. In the proposed pruned git
repository, these are the largest files.

This standardizes both remaining test profile directories on the same set of
binaries (they're currently out of sync), and moves them to new path names so
we can purge the previous paths from history.

BUG=111570
TEST=none

Review URL: https://chromiumcodereview.appspot.com/10698009

TBR=erg@chromium.org

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@153952 0039d316-1c4b-4281-b951-d872f2087c98