Convert the history pyauto tests to browser_tests.
testStarredBookmarkInHistory was skipped since it wasn't very useful (it got the star state from the bookmark code, which it also called)
testForge is testing adding and reading a url from the history database, which is a duplicate of many browser and unit tests (i.e. HistoryTest.*)
testHttpsHistory is a duplicate of HistoryTest.MakeIntranetURLsTyped
testFtpHistory is not very useful since if multiple schemes work, we don't need to test every scheme
testAppendTabAddPage is a duplicate of a subset of testReloadBringPageToTop
testOpenWindowAddPage is a duplicate of a subset of testMultiTabsWindowsHistory

I fixed a bug in ui_test_utils::NavigateToURLWithDisposition which made it not wait on NEW_BACKGROUND_TAB disposition. I also made HistoryEnumerator return the urls in sorted order.

BUG=143637
Review URL: https://chromiumcodereview.appspot.com/10915002

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@154229 0039d316-1c4b-4281-b951-d872f2087c98