webui: Request favicons at all required display scale factors
Note: This doesn't help for webui pages that set the favicon
to a lodpi png using <link rel=icon>. The options page does
this to switch the favicon for different views.

For this reaosn, remove an uneeded link rel=icon from the
downloads page.

BUG=145228

Review URL: https://chromiumcodereview.appspot.com/10908096

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@154991 0039d316-1c4b-4281-b951-d872f2087c98