Remove cache of XRandR data and be stateless.
Since the multi-monitor support code was first implemented back in powerd, we
have always had rare and unrecreatable bugs where we see inconsistent XRandR
data.  The prime suspect of this problem is our reliance on the XRandR cache
which has existed in some form since the earliest versions of this code and
has always had the potential problem of allowing XRandR data to escape the
lifespan of the event being processed.
Additionally, bugs are periodically introduced due to needing to keep this cache
in sync with the underlying X server process.
By making OutputConfigurator stateless,  the potential of a stale or otherwise
inconsistent cache disappears.

BUG=chromium:142860
TEST=Manually tested on Lumpy:  add/remove second display, cycle displays,
manually power down and up the internal panel.


Review URL: https://chromiumcodereview.appspot.com/10907078

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@155211 0039d316-1c4b-4281-b951-d872f2087c98