Pass Me2Me config via stdin.
This allows the host to restart even if the config file is no longer readable (for example, if it is on an NFS share). One downside is that the host can no longer re-read its configuration without restarting.

I ended up implementing a SIGTERM handler to provide clean shutdown. It's easily extractable into a separate CL; let me know if you'd like me to do that.

BUG=None


Review URL: https://chromiumcodereview.appspot.com/10905081

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@155314 0039d316-1c4b-4281-b951-d872f2087c98