drive: Get rid of DocumentEntry::GetAllEntryKinds()
DocumentEntry::GetAllEntryKinds() was necessary as EntryKind enum
was using complex bitmasks, rather than consecutive numbers.
This was fragile as we needed to keep EntryKind and EntryKindMap
in sync manually without help from the compiler.

In this patch, EntryKind is changed to consecutive numbers,
ClassifyEntryKind() is introduced, and GetAllEntryKinds() is
removed.

BUG=135162
TEST=unit_tests

Review URL: https://chromiumcodereview.appspot.com/10912160

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@155519 0039d316-1c4b-4281-b951-d872f2087c98