Fix DB migration for upgrading to version 22 for Android
When upgrading to version 22, there is an issue for Android, where a
table is dropped and recreated with indexes. However, since the indexes
were not dropped, they would fail on creation.

This adds the creation of the indexes to the unit test SQL commands, and
fixes the issue by dropping the indexes from the DB before recreating
the table and its indexes.

BUG=None


Review URL: https://chromiumcodereview.appspot.com/10912153

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@155545 0039d316-1c4b-4281-b951-d872f2087c98