Add is_fullscreen to video updates
This is a re-attempt at http://codereview.chromium.org/10916123/, which also ended up being reverted.
    
This was reverted due to a failing unit test on Win_Aura. Through an unfortunate set of mistakes this builder was not in the default try set or the CQ set, but could close the tree leading to the revert. Through testing I have been able to reproduce the failure on linux_chromeos, so it might be a flakey failure, since it passed sometimes on that bot. This CL is basically the same as the CL list above, except that there is now a window->Focus() call in the unit test, since the reason the test was failing was when it was being made full screen it ceased to visible. This fixes the issue within this test. I will added a comment about this on the bug tracking the underlying issue.
    
This is a re-attempt at http://codereview.chromium.org/10905026/, which ended up being reverted.
    
The crash that caused the revert was actually due to powerd not being able to parse the protobuf correctly. https://gerrit.chromium.org/gerrit/32243/ corrects this issue, so once that CL lands this CL is good to go. I confirmed the combination of CLs works correctly.
    
This is the original CL description:
    
This adds a boolean to the video update message sent to powerd that indicates whether or not the video playing window is fullscreen'd. This is information used in powerd for controlling when to enable/disable keyboard backlights. Specifically if video is playing and fullscreen we assume that the user is doing something like watching a movie so will disable backlight.
    
An additional change that I have made is to convert the message being sent from an int64 to a protobuffer that contains a int64 and a boolean. Currently, since the receiver of this message does not live in the same repo it means that every time one wants to change the message format there is a small window where bad builds can be generated or one has to land two changes to the other repo, one to add handling of the new message format and then a second one to remove the old format. Converting to a protobuffer means that for common cases this issues are removed.
    
This CL depends on CLs for power_manager (https://gerrit.chromium.org/gerrit/32092/), one for system_api (https://gerrit.chromium.org/gerrit/31916), and a DEPS roll(http://codereview.chromium.org/10915032/).  There will be future CLs for power_manager that will use this information and more thourghly tests this code.
    
BUG=chrome-os-partner:9203
TEST=Run updated unittests.
     Confirm there is no messages in the cros logs about this method being mishandled.
     Make sure that powerd is not crashing due to protocol buffer processing failing.


Review URL: https://chromiumcodereview.appspot.com/10913134

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@156008 0039d316-1c4b-4281-b951-d872f2087c98