Shorten the path length to get under MAX_PATH on windows.
Some of the nested directories in each partition have variable length path names. Prepending our isolated storage path structure used to take us over the to MAX_PATH for chrome-extension:// origins when creating the database files for subsystems like Local Storage.

With this change, on XP, assuming a 10-character user name, the Local Storage database file representing the chrome-extension: origin in the Default profile for user "abcdefghij" has this path:

 C:\Documents and Settings\abcedfghij\Local Settings\Application Data\
 Google\Chrome SxS\User Data\Default\Storage\
 ext\bdlahefabekefwlefaancobndodocndn\def\Local Storage\
 chrome-extension_bdlahefabekefwlefaancobndodocndn_0.localstorage-journal

which is 241 characters. This gives 19 characters of headroom from the 260 MAX_PATH.

Previously, when we used "Storage Partitions", "extensions", and "default" instead of "Storage, "ext", and "def", the same database file would have a path length of 263 characters.

This change doesn't completely solve the problem.  If Local Storage is used on a super-long domain, we can still exceed MAX_PATH.  However, using my own profile as an example, of 342 domains, none created paths longer than those of the chrome-extensions so this should be a pretty solid mitigation.

BUG=151450

Review URL: https://codereview.chromium.org/10967030

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@158136 0039d316-1c4b-4281-b951-d872f2087c98